Re-write for refiller module
Use: place in chassie on any side of an apiary/alveary.  Module will
constantly check if the queen has died.  When she has, it will check
all slots for the offspring princess and put her back into top breeding
slot.  Then it will do the same for a drone.  It does not differentiate
between drones YET.  I may make it only select drones that are the same
species as the queen that died. (This way, it wont mess up mutations.)

This will ALSO extract any products AUTOMATICALLY as they are produced.