don't do linear searches over all nbt tags!
changed to be a concurrent map<itemKey,map<tags,int> so that it is easy
to find if something exists, and what its tag is.
can't do a direct itemIdentifier, because at the early stage it we don't
know the itemID, which is what is used for the itemIdentifier hash.