Items can not deliver to the inventory they are pulling from
before a block can deliver, the request manager checks that the source
IInventory isn't the destination IInventory.

This will break cases where an ISided inventory is both the source and
the destination for the same thing. this will require another chest to
buffer the requests.

This does allow buffers for crafting networks, by having a supplier and
a provider on the same chest, with 16 of something requested.

The next commit will cause ACT's to drain their buffers, not requesting
matierals if they are already in teh ACT or assocaited chest.