use the gregtech api to avoid reflection
i do not know what will happen if more than 1 stack/tick is tried to be
removed. Previously we pulled from the internal variable for the amount
that was there. I expect that broke stuff in GregTech when the internal
var for the amount was not the same as the internal stack size. A recent
GT patch broke our old code, see #555