Rewrote internal Item handling inside LP pipes.
We now have our own ItemTransportation code which uses progress percentage instead of position coordinates.
That allows the use of the partial render tick to be used to render item transportation more smoothly.
The Routing information are now stored inside an additional class which than can be linked to TravelingItems from other transportation systems such as BC or TE.