Allow LP to deal with recipe overlapping.
Allows users do decide which version of for example a item they want to have. Either the TE version or the IC2 version.
This also helps if there is a recipe overlapping another with a completely different outcome. With this the user can bypass that problem until it is fixed by the original mod author.