handle partial insert into a full InvSysConnecter
LPTravelingItemServer holds a stack.
reverseItem takes only a LPTIS, and not an extra stack, which is a clone
from it.
Added comments to indicate which functions take ownership of the stack
passed into it.

itemStack clone of ArrivingItem removed from code, as nothing takes
ownership of it, and passing around things which may or may not be
duplicate or different is just asking for trouble.

when we call InsertedItemStack on a partial insert, duplicate the info
and notify of insert of the partial stack, bounce the rest.

inserted does not take a redundant item and info, only an info (updated
to hold the correct stack size)

handleItemEnterInv removes old addItem code

PipeItemsInvSysConnector::
addItem now just takes an info, specifying both the item and the
destination.
addItem now ignores 0 size stacks.

handleItemEnterInv now only takes an info and a tile.

IDirectRoutingConnection updated to accoutnt for changed addItem
changes.