Update StorageDrawers API so dev environment works
The StorageDrawers(SD) API has changed a bit from the version we have in
LP.
This causes the dev environment to crash upon ticking of a SD drawer
controller.

This change is not needed to fix any actual bugs in LP, as the compiled
jar works fine, but this change prevents the Dev environment from
crashing.