Start making an enhanced guide book.
Added simple Item, still have to attach it to the event listener and add a recipe but that's not a priority.
Added a scale-able GUI that works around the already used translation based book text.
Semi-implemented slider functionality but that's also not a priority as of now.

Reworked background drawing
All basic functionality of a normal book (and more) are implemented.