Fix #1373
Prevent crash when an item is inside an InvSysCon going to an inventory and the server is restarted. In hat case the item does not have a destination anymore, but it is already heading to the inventory so in that case we just ignore it from an InvSysCon point of view to prevent a crash.