Multiple GL11 flag resets. Fix #1441
Reset color and flags with GlStateManager and GL11#glEnable/glDisable in multiple points in  both the base LogisticsBaseGuiScreen.java class as well as the GuiCraftingPipe.java class to atone for some GlStateManager leaks that have yet to be identified.
This will fix the current problem but this is not a complete fix, this is just a simple hack.