Create first async helpers
AsyncRouting has access to the ServerRouter and provides a update routing
table coroutine as well as direct access to the ServerRouter.

AsyncLogisticsManager currently implements the getDestination function from
the LogisticsManager with some boost on the getBestReply function as well as
an allDestinations function that returns a Kotlin Flow. Flows are based on
the Reactive Streams concept.