Added new BookContent object class.
Added new BookContent class that will be responsible for the guidebook parsing.
All the classes that need to interact with it were refactored and are now all working fine together.
Some of the refactored classes have been moved into the rs845 package and are now under the MIT license.
Some classes just had to have the imports updated.
Unused classes were removed.