Tweaked the BookContents and page structure
The Yaml information on the pages will now be a Map with string keys and Map values.
Each entry will represent a possible button set on the page. Each entry's map is the just like it used to be, string keys and list of strings for values.
In the markdown part of the page there will be a markdown entry where you need to specify the first level key of the map and the kind button set it will be (tiles or list).
It should look something like this for example: $[map_key](type=menu)