Created button superclass which fixed multiple issues.
All buttons on the ui now extend LPGuiButton.kt (name subject to change) and so most shared functionality is now implemented in the superclass. This fixed the weird "hitboxes", misclicks and in general cleared the code quite a bit.
The button that adds bookmarks now doubles as the button that removes bookmarks, meaning it's a stated button.