Fixed element padding
Removed individual element padding and added padding in between drawable paragraphs on setPos.
Each paragraph should only add padding in between it's on own children, not before or after.
An exception to this rule is made with the HorizontalLineParagraph because it's whole point is to separate other paragraphs and so it makes sense that it would add padding, but only after itself.